% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk.FactorAnalysis-package.R
\docType{package}
\name{rk.FactorAnalysis-package}
\alias{rk.FactorAnalysis}
\alias{rk.FactorAnalysis-package}
\title{\packageTitle{rk.FactorAnalysis}}
\description{
\packageDescription{rk.FactorAnalysis}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab rk.FactorAnalysis\cr
Type: \tab Package\cr
Version: \tab 0.2-1\cr
Date: \tab 2022-05-03\cr
Depends: \tab rkward (>= 0.6.5)\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://rkward.kde.org\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rkward.kde.org}
}

}
\author{
\packageAuthor{rk.FactorAnalysis}

Maintainer: \packageMaintainer{rk.FactorAnalysis}
}
