\docType{package}
\name{rk.FactorAnalysis-package}
\alias{rk.FactorAnalysis}
\alias{rk.FactorAnalysis-package}
\title{The rk.FactorAnalysis Package}
\description{
  RKWard GUI to conduct principal component and factor
  analysis.
}
\details{
  \tabular{ll}{ Package: \tab rk.FactorAnalysis\cr Type:
  \tab Package\cr Version: \tab 0.01-10\cr Date: \tab
  2013-02-24\cr Depends: \tab rkward (>= 0.5.6)\cr
  Enhances: \tab rkward\cr License: \tab GPL (>= 3)\cr
  LazyLoad: \tab yes\cr URL: \tab http://rkward.sf.net\cr }

  RKWard GUI to conduct principal component and factor
  analysis (using the psych package). Also includes dialogs
  for scree plots, correlation plots, VSS/MAP and parallel
  analysis.
}
\author{
  Meik Michalke \email{meik.michalke@hhu.de}
}
\keyword{package}

