% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rk.FactorAnalysis-package.R
\docType{package}
\name{rk.FactorAnalysis-package}
\alias{rk.FactorAnalysis-package}
\title{The rk.FactorAnalysis Package}
\description{
RKWard GUI to conduct principal component and factor analysis.
}
\details{
\tabular{ll}{
Package: \tab rk.FactorAnalysis\cr
Type: \tab Package\cr
Version: \tab 0.01-13\cr
Date: \tab 2015-03-14\cr
Depends: \tab rkward (>= 0.6.0)\cr
Enhances: \tab rkward\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://rkward.kde.org\cr
}

RKWard GUI to conduct principal component and factor analysis (using the psych package). Also includes dialogs for scree plots,
      correlation plots, VSS/MAP and parallel analysis.
}
\author{
Meik Michalke
}
\keyword{package}

